Dev adam xla and rm sys op (#3584)
* rm model_update sys op

* mv indexed slice op kernel to model_update op

* adam optimizer

* revert indexed_slices_reduce_sum op

* fix

* refine

* rm mask in lazy_adam testcase

* add cast
