+++ oneflow/core/job_rewriter/optimizer.cpp
@@ -105,11 +105,7 @@ void ConstructMdUpdtOpConf(const VariableOp& op, const LogicalBlobId& diff_lbi_o
                                          const LogicalBlobId& diff_lbi_of_var_out, \
                                          JobBuilder* job_builder, T* mdupdt_op_conf)
 
-INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(NaiveModelUpdateOpConf);
-INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(MomentumModelUpdateOpConf);
 INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(RMSPropModelUpdateOpConf);
 INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(LARSModelUpdateOpConf);
-INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(AdamModelUpdateOpConf);
-INSTANTIATE_CONSTRUCTOR_MDUPDT_OP_CONF(LazyAdamModelUpdateOpConf);
 
 }  // namespace oneflow
