+++ oneflow/core/kernel/kernel.proto
@@ -68,24 +68,6 @@ message IndexedSlicesReduceSumKernelConf {
   required DataType indices_data_type = 1;
 }
 
-message IndexedSlicesNaiveModelUpdateKernelConf {
-  required DataType indices_data_type = 1;
-  required int64 lower_bound = 2;
-  required int64 upper_bound = 3;
-}
-
-message IndexedSlicesMomentumModelUpdateKernelConf {
-  required DataType indices_data_type = 1;
-  required int64 lower_bound = 2;
-  required int64 upper_bound = 3;
-}
-
-message IndexedSlicesLazyAdamModelUpdateKernelConf {
-  required DataType indices_data_type = 1;
-  required int64 lower_bound = 2;
-  required int64 upper_bound = 3;
-}
-
 message SyncDynamicResizeKernelConf {
   required DataType size_data_type = 1;
 }
@@ -114,9 +96,6 @@ message KernelConf {
     XrtLaunchKernelConf xrt_launch_conf = 353;
     UniqueWithCountsKernelConf unique_with_counts_conf = 354;
     IndexedSlicesReduceSumKernelConf indexed_slices_reduce_sum_conf = 355;
-    IndexedSlicesNaiveModelUpdateKernelConf indexed_slices_naive_model_update_conf = 357;
-    IndexedSlicesMomentumModelUpdateKernelConf indexed_slices_momentum_model_update_conf = 358;
-    IndexedSlicesLazyAdamModelUpdateKernelConf indexed_slices_lazy_adam_model_update_conf = 359;
     SyncDynamicResizeKernelConf sync_dynamic_resize_conf = 360;
     ArgWhereKernelConf arg_where_conf = 361;
