+++ oneflow/core/operator/op_conf.proto
@@ -353,25 +353,6 @@ message NormalModelUpdateOpConf {
   optional float weight_decay = 7 [default = 0.0];
 }
 
-message NaiveModelUpdateOpConf {
-  required NormalModelUpdateOpUserConf user_conf = 1;
-  required string model_diff = 2;
-  required string model = 4;
-  required string train_step = 5;
-  required string learning_rate = 6;
-  optional float weight_decay = 7 [default = 0.0];
-}
-
-message MomentumModelUpdateOpConf {
-  required NormalModelUpdateOpUserConf user_conf = 1;
-  required string momentum = 2;
-  required string model_diff = 3;
-  required string model = 5;
-  required string train_step = 6;
-  required string learning_rate = 7;
-  optional float weight_decay = 8 [default = 0.0];
-}
-
 message RMSPropModelUpdateOpConf {
   required NormalModelUpdateOpUserConf user_conf = 1;
   required string model_diff = 2;
@@ -391,32 +372,6 @@ message LARSModelUpdateOpConf {
   required string train_step = 6;
   required string learning_rate = 7;
   optional float weight_decay = 8 [default = 0.0];
-}
-
-message AdamModelUpdateOpConf {
-  required NormalModelUpdateOpUserConf user_conf = 1;
-  required string m = 2;
-  required string v = 3;
-  optional string beta1_t = 4;
-  optional string beta2_t = 5;
-  required string model_diff = 6;
-  required string model = 8;
-  required string train_step = 9;
-  required string learning_rate = 10;
-  optional float weight_decay = 11 [default = 0.0];
-}
-
-message LazyAdamModelUpdateOpConf {
-  required NormalModelUpdateOpUserConf user_conf = 1;
-  required string m = 2;
-  required string v = 3;
-  optional string beta1_t = 4;
-  optional string beta2_t = 5;
-  required string model_diff = 6;
-  required string model = 8;
-  required string train_step = 9;
-  required string learning_rate = 10;
-  optional float weight_decay = 11 [default = 0.0];
 }
 
 message AccumulateOpConf {
@@ -886,36 +841,6 @@ message ConstantLikeOpConf {
   }
 }
 
-message IndexedSlicesNaiveModelUpdateOpConf {
-  required string model_diff_indices = 1;
-  required string model_diff_values = 2;
-  required string model = 3;
-  required string learning_rate = 4;
-}
-
-message IndexedSlicesMomentumModelUpdateOpConf {
-  required string momentum = 1;
-  required string model_diff_indices = 2;
-  required string model_diff_values = 3;
-  required string model = 4;
-  required string train_step = 5;
-  required string learning_rate = 6;
-  optional float beta = 7 [default = 0.9];
-}
-
-message IndexedSlicesLazyAdamModelUpdateOpConf {
-  required string m = 1;
-  required string v = 2;
-  required string model_diff_indices = 3;
-  required string model_diff_values = 4;
-  required string model = 5;
-  required string train_step = 6;
-  required string learning_rate = 7;
-  optional float beta1 = 8 [default = 0.9];
-  optional float beta2 = 9 [default = 0.999];
-  optional float epsilon = 10 [default = 1e-8];
-}
-
 message SyncDynamicResizeOpConf {
   required string in = 1;
   required string size = 2;
@@ -988,11 +913,8 @@ message OperatorConf {
     TickOpConf tick_conf = 124;
     KeepHeaderOnlyOpConf keep_header_only_conf = 125;
     TotalLossInstanceNumOpConf total_loss_instance_num_conf = 126;
-    NaiveModelUpdateOpConf naive_model_update_conf = 127;
-    MomentumModelUpdateOpConf momentum_model_update_conf = 128;
     RMSPropModelUpdateOpConf rmsprop_model_update_conf = 129;
     LARSModelUpdateOpConf lars_model_update_conf = 130;
-    AdamModelUpdateOpConf adam_model_update_conf = 131;
     ShapeElemCntOpConf shape_elem_cnt_conf = 132;
     AccOpConf acc_conf = 133;
     SourceTickOpConf source_tick_conf = 135;
@@ -1007,7 +929,6 @@ message OperatorConf {
     AccTickOpConf acc_tick_conf = 144;
     ReturnOpConf return_conf = 146;
     PartialTickOpConf partial_tick_conf = 149;
-    LazyAdamModelUpdateOpConf lazy_adam_model_update_conf = 150;
     ForeignWatchOpConf foreign_watch_conf = 151;
     DistributeConcatOpConf distribute_concat_conf = 155;
     DistributeSplitOpConf distribute_split_conf = 156;
@@ -1016,9 +937,6 @@ message OperatorConf {
     DeviceTickOpConf device_tick_conf = 159;
     DynamicBinarySplitOpConf dynamic_binary_split_conf = 160;
     DynamicBinaryConcatOpConf dynamic_binary_concat_conf = 161;
-    IndexedSlicesNaiveModelUpdateOpConf indexed_slices_naive_model_update_conf = 163;
-    IndexedSlicesMomentumModelUpdateOpConf indexed_slices_momentum_model_update_conf = 164;
-    IndexedSlicesLazyAdamModelUpdateOpConf indexed_slices_lazy_adam_model_update_conf = 165;
     SliceBoxingCopyOpConf slice_boxing_copy_conf = 166;
     SliceBoxingAddOpConf slice_boxing_add_conf = 167;
     TensorBufferToTensorListOpConf tensor_buffer_to_tensor_list_conf = 168;
